Remove boilerplate code
